mcp-server/requirements.txt
# Core MCP dependencies
mcp>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0

# For async operations
aiohttp>=3.9.0
